Rename playlist widgets
